+++ main1034.cpp
@@ -260,7 +260,7 @@ case 1:
 SSq4 = kNoFKBjt((unsigned short)-527284668, -534235493);
 break;
 case 2:
-VJ2Ec15Q8C = xQi46((unsigned short)1084590885, C0oq4x, (short)-1605676254, i5IW0y, -1761235472);
+VJ2Ec15Q8C = xQi46((unsigned short)1084590885, C0oq4x, (short)-1605676254, i5IW0y, (short)-1761235472);
 break;
 case 3:
 EO4azv = kNoFKBjt(-2086180096, 133220896);
