+++ main1129.cpp
@@ -16,7 +16,7 @@ int main(void) {
     KE7720AOK    =    (short)-794024917;
     tK_v89    =    (short)421339480;
     vM0DqmJH    =    (short)-138108378;
-    Oi6sDe9x    =    1432181015;
+    Oi6sDe9x    =    (short)1432181015;
     gLNlp    =    1434747066;
     pfVDzc8    =    1591850993;
     n_Y6IKRZ    =    2136730805;
