+++ main1139.cpp
@@ -26,7 +26,7 @@ int main() {
 	c020UceE	=	(unsigned short)1376937137;
 	D9xd2292HN	=	(unsigned short)-823020540;
 	jyhnie7u9	=	(unsigned short)-334273957;
-	JD058e2	=	-1146721985;
+	JD058e2	=	(unsigned short)-1146721985;
 	qOn3hb	=	-572551953;
 	Tec84Bfcw	=	882317617;
 	qM4d	=	1630010055;
@@ -36,7 +36,7 @@ int main() {
 	FXmlDg_Be	=	(unsigned short)755374072;
 	prKswuR0F	=	(unsigned short)891265006;
 	V55OC_2Ebq	=	(unsigned short)-995803375;
-	v0jbiRiTi	=	-9811487;
+	v0jbiRiTi	=	(unsigned short)-9811487;
 	y6MNC5uY2	=	-520829028;
 	e_DSS	=	687839690;
 	rb198R04	=	-1209714125;
@@ -46,7 +46,7 @@ int main() {
 	dqWHh	=	(unsigned short)-1543255077;
 	KWh5m0	=	(unsigned short)-2054403666;
 	Hh1Fm4_	=	(unsigned short)2034122852;
-	RB_5	=	-111721498;
+	RB_5	=	(unsigned short)-111721498;
 	a9A095fDiD	=	49658445;
 	NkZw4mgp	=	1822202263;
 	V_0877	=	-379864093;
@@ -56,7 +56,7 @@ int main() {
 	ISV_AdmC5p	=	(unsigned short)1350523359;
 	LyZhPbd1	=	(unsigned short)480294482;
 	Gc_lFC69	=	(unsigned short)-1587994485;
-	UBQkK	=	-559014399;
+	UBQkK	=	(unsigned short)-559014399;
 	rTSQMG9jc	=	-123987900;
 	HisuZ1Tn	=	761447537;
 	lgH7FK	=	-2026957172;
@@ -66,7 +66,7 @@ int main() {
 	h469Vx4	=	(unsigned short)-1928217296;
 	LGjz8	=	(unsigned short)740461384;
 	WcRn	=	(unsigned short)-103602187;
-	eri4N7ko	=	-1186899969;
+	eri4N7ko	=	(unsigned short)-1186899969;
 	SQI2g0lS8	=	-1734199649;
 	Z76k_A2J_	=	346730950;
 	T0gzp7bnl	=	-2006039250;
@@ -76,7 +76,7 @@ int main() {
 	tU4_i	=	(unsigned short)-435986848;
 	itS1SM7b	=	(unsigned short)-1178059269;
 	hkha1	=	(unsigned short)-1625717713;
-	NCO9Rz7RDb	=	-647662377;
+	NCO9Rz7RDb	=	(unsigned short)-647662377;
 	jBFx	=	-1839163037;
 	JrIvSreY	=	-754685068;
 	ZshXX_	=	-760534980;
@@ -86,7 +86,7 @@ int main() {
 	IUMAcw0	=	(unsigned short)-1871188828;
 	pOQEVa	=	(unsigned short)-2006134477;
 	IGoZUCd	=	(unsigned short)-61552867;
-	eJE63_b	=	326305653;
+	eJE63_b	=	(unsigned short)326305653;
 	VCuM3g	=	2033577319;
 	g4b7l3E3	=	-1918695512;
 	cdpA3vDt9	=	859574619;
